send multiple fake pms
